+++ test/selectors/extractor-test.js
@@ -1,6 +1,6 @@
 var vows = require('vows');
 var assert = require('assert');
-var tokenize = require('../../lib/selectors/tokenizer');
+var tokenize = require('../../lib/tokenizer/tokenize');
 var extractor = require('../../lib/selectors/extractor');
 
 function buildToken(source) {